refactor: IPNE 描画のオーバーレイ層を drawOverlays へ分離
- 低HP警告/コンボ/ボスWARNING/シェイク復元/暗転/ステージ演出/自動マップ/デバッグを逐語移植
- シェイク restore の位置と bossWarningRef 等の副作用を保存
- renderGameFrame はセットアップ＋5層呼び出しのみに。特性化テストで不変を確認
